+++ src/transformer/index.ts
@@ -31,6 +31,7 @@ import { rtHelper } from './rt-helper';
 import { serialize } from './serialize';
 import * as ts from 'typescript';
 import { cloneEntityNameAsExpr, getRootNameOfEntityName } from './utils';
+import { T_ANY, T_ARRAY, T_INTERSECTION, T_TUPLE, T_UNION, T_UNKNOWN } from '../common';
 
 export enum TypeReferenceSerializationKind {
     // The TypeReferenceNode could not be resolved.
@@ -241,33 +242,50 @@ const transformer: (program : ts.Program) => ts.TransformerFactory<ts.SourceFile
                 else if (typeNode.kind === ts.SyntaxKind.BigIntKeyword)
                     return ts.factory.createIdentifier('BigInt');
                 else if (typeNode.kind === ts.SyntaxKind.AnyKeyword)
-                    return ts.factory.createIdentifier('Object');
+                    return serialize({ TΦ: T_ANY });
                 else if (typeNode.kind === ts.SyntaxKind.FunctionType)
                     return ts.factory.createIdentifier('Function');
                 else if (typeNode.kind === ts.SyntaxKind.UnknownKeyword)
-                    return ts.factory.createIdentifier('Object');
+                    return serialize({ TΦ: T_UNKNOWN });
                 else if (ts.isArrayTypeNode(typeNode)) {
                     if (extended)
-                        return ts.factory.createArrayLiteralExpression([serializeTypeRef(typeNode.elementType, true)]);
+                        return serialize({ TΦ: T_ARRAY, e: literalNode(serializeTypeRef(typeNode.elementType, true)) });
                     else
                         return ts.factory.createIdentifier('Array');
                 }
                 
+                if (ts.isTupleTypeNode(typeNode)) {
+                    if (!extended)
+                        return ts.factory.createIdentifier('Object');
+                    
+                    return serialize({
+                        TΦ: T_TUPLE,
+                        e: typeNode.elements.map(e => {
+                            if (ts.isNamedTupleMember(e)) {
+                                return { n: e.name.text, t: literalNode(serializeTypeRef(e.type, extended)) };
+                            } else {
+                                return { t: literalNode(serializeTypeRef(e, extended)) };
+                            }
+                        })
+                    })
+                    
+                }
+
                 if (ts.isUnionTypeNode(typeNode)) {
                     if (!extended)
                         return ts.factory.createIdentifier('Object');
 
                     return serialize({
-                        kind: 'union',
-                        types: typeNode.types.map(x => literalNode(serializeTypeRef(x, extended)))
+                        TΦ: T_UNION,
+                        t: typeNode.types.map(x => literalNode(serializeTypeRef(x, extended)))
                     });
                 } else if (ts.isIntersectionTypeNode(typeNode)) {
                     if (!extended)
                         return ts.factory.createIdentifier('Object');
                     
                     return serialize({
-                        kind: 'intersection',
-                        types: typeNode.types.map(x => literalNode(serializeTypeRef(x, extended)))
+                        TΦ: T_INTERSECTION,
+                        t: typeNode.types.map(x => literalNode(serializeTypeRef(x, extended)))
                     });
                 }